moved pet related constants to a new file
