+++ src/data/constants.ts
@@ -27,17 +27,3 @@ export const basicOptions: string[] = [
 export const defaultDate: Date = new Date(0);
 
 export const todayDate: Date = new Date();
-
-export const animalGroupOptions: string[] = [
-    "Small mammal (e.g. cat, rabbit, and mouse)",
-    "Farm (e.g. chicken, pig, and cow)",
-    "Equine (e.g. horse)",
-    "Bird",
-    "Reptile",
-    "Amphibian",
-    "Fish",
-    "Invertebrate",
-    "Other",
-];
-
-export const sexOptions: string[] = ["Male", "Female", "Unknown"];
