+++ src/data/pets/constants.ts
@@ -1,0 +1,13 @@
+export const animalGroupOptions: string[] = [
+    "Small mammal (e.g. cat, rabbit, and mouse)",
+    "Farm (e.g. chicken, pig, and cow)",
+    "Equine (e.g. horse)",
+    "Bird",
+    "Reptile",
+    "Amphibian",
+    "Fish",
+    "Invertebrate",
+    "Other",
+];
+
+export const sexOptions: string[] = ["Male", "Female", "Unknown"];
